Rates and sum calculation on main screen

+++ app/src/main/java/ru/pavelcoder/modulbankdemo/model/bank/CurrencyRate.kt
@@ -3,5 +3,5 @@ package ru.pavelcoder.modulbankdemo.model.bank
 data class CurrencyRate (
     val base: Currency,
     val target: Currency,
-    val rate: Double
+    val rate: Float
 )
